Add useList hook
